writing geometry stats now. need to verirfy they are correct however
