+++ CMakeLists.txt
@@ -97,6 +97,7 @@ set(EXTENSION_SOURCES
 	src/planning/metadata_io/manifest_list/iceberg_manifest_list_reader.cpp
 	src/planning/pruning/iceberg_predicate.cpp
 	src/planning/snapshot/iceberg_snapshot_lookup.cpp
+	src/storage/statistics/iceberg_statistics.cpp
 )
 
 add_subdirectory(src/rest_catalog/objects)
